+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed  
 
--
+- Only print flow lines in 'Query by Fire Perimeters' workflow
 
 ### Deprecated